# openapi-spec.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/v1/invoice/invoices/{id}": {
            "get": {
                "tags": [
                    "invoice-controller"
                ],
                "operationId": "getInvoiceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Invoice"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "invoice-controller"
                ],
                "operationId": "updateInvoice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Invoice"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Invoice"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "invoice-controller"
                ],
                "operationId": "deleteInvoice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/v1/invoice/addInvoices": {
            "post": {
                "tags": [
                    "invoice-controller"
                ],
                "operationId": "createInvoice",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Invoice"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/adjustment/addAdjustment": {
            "post": {
                "tags": [
                    "adjustment-controller"
                ],
                "operationId": "createInvoice_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Adjustment"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Adjustment"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/invoice/invoicesForMonth": {
            "get": {
                "tags": [
                    "invoice-controller"
                ],
                "operationId": "getInvoiceForMonth",
                "parameters": [
                    {
                        "name": "month",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Invoice"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/invoice/invoicesForMonthAndYear": {
            "get": {
                "tags": [
                    "invoice-controller"
                ],
                "operationId": "getInvoiceForMonthAndYear",
                "parameters": [
                    {
                        "name": "selectedDate",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Invoice"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/invoice/getInvoicesForProject": {
            "get": {
                "tags": [
                    "invoice-controller"
                ],
                "operationId": "getInvoicesForProject",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Invoice"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/invoice/getInvoicesForEmployee": {
            "get": {
                "tags": [
                    "invoice-controller"
                ],
                "operationId": "getInvoicesForEmployee",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Invoice"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/invoice/activeInvoices/{monthYear}": {
            "get": {
                "tags": [
                    "invoice-controller"
                ],
                "operationId": "getActiveInvoicesForMonth",
                "parameters": [
                    {
                        "name": "monthYear",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Invoice"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Invoice": {
                "type": "object",
                "properties": {
                    "invoiceId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "projectId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "invoiceMonth": {
                        "type": "string",
                        "format": "date"
                    },
                    "billing": {
                        "type": "number",
                        "format": "float"
                    },
                    "hours": {
                        "type": "number",
                        "format": "float"
                    },
                    "total": {
                        "type": "number",
                        "format": "float"
                    },
                    "invoicePaidAmount": {
                        "type": "number",
                        "format": "float"
                    },
                    "invoiceDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "paymentDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "type": "string"
                    }
                }
            },
...
            "Adjustment": {
                "type": "object",
                "properties": {
                    "adjustmentId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "fromId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "toId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "amount": {
                        "type": "number",
                        "format": "double"
                    },
                    "adjustmentType": {
                        "type": "string"
                    },
                    "notes": {
                        "type": "string"
                    },
                    "adjustmentDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
...
        }
    }
}